Tutorial
I have changed the images to the prefab but I dont like the animation (they roll). I have put also a button to Luca's part which should appear in olly's third audio. I have also put the function to switch the camera but I was thinking in changing the scene to Luca's scene as well. Maybe we can ask the kid at the end of the tutorial if they want to listen again the tutorial or continue. If he ask again for the tutorial the scene just need to be loaded again, either way we can continue to Giuseppe's scene